+++ build.sh
@@ -1,1 +1,1 @@
-gcc src/*.c CAE/include/CAE.c CAE/include/b64/*.c -o build/letonia -lallegro -lallegro_primitives -lallegro_image -lm -lallegro_ttf -lallegro_font -lallegro_audio -lallegro_acodec -lallegro_video
+gcc src/*.c src/scenes/*.c CAE/include/CAE.c CAE/include/b64/*.c -o build/letonia -lallegro -lallegro_primitives -lallegro_image -lm -lallegro_ttf -lallegro_font -lallegro_audio -lallegro_acodec -lallegro_video
