+++ src/scenes/roomLeft.c
@@ -1,0 +1,37 @@
+#include "../../include/scenes.h"
+
+void loadRoomLeft(){
+    Scene* roomL = createScene(engine, gameSceneScript);
+    roomL->camera.minLimit.x = 0;
+    roomL->camera.minLimit.y = 0;
+    roomL->camera.maxLimit.x = 1001;
+    roomL->camera.maxLimit.y = 501;
+    roomL->camera.zoom = 2.5;
+    roomL->camera.followTarget = player;
+    setupSceneWorld(roomL, loadBitmap(engine, "./assets/images/roomL.png"), 500, 500);
+    addWorldTile(roomL, 0, 0, 0, 0);
+    addGameObjectToScene(roomL, player);
+    roomLobj = createGameObject(SOLID, 12, 12, 448, 468, roomL);
+    roomLobj->color = al_map_rgba(0, 0, 0, 0);
+    roomLobj->collisionEnabled = 1;
+    roomLobj->collisionType = COLLISION_RECT;
+    roomLobj->invertedCollision = 1;
+
+    GameObject* fundoRoomL = createGameObject(SOLID, 500, 0, 1080, 1080, roomL);
+    fundoRoomL->color = al_map_rgba(184, 118, 93, 255);
+
+    GameObject* tableRoomL = createGameObject(SOLID, 70, 68, 260, 90, roomL);
+    tableRoomL->color = al_map_rgba(0, 0, 0, 0);
+    tableRoomL->collisionEnabled = 1;
+    tableRoomL->collisionType = COLLISION_RECT;
+
+    GameObject* benchRoomL = createGameObject(SOLID, 138, 25, 125, 28, roomL);
+    benchRoomL->color = al_map_rgba(0, 0, 0, 0);
+    benchRoomL->collisionEnabled = 1;
+    benchRoomL->collisionType = COLLISION_RECT;
+
+    returnQuartelRoomL = createGameObject(SOLID, 30, 478, 208, 28, roomL);
+    returnQuartelRoomL->color = al_map_rgba(0, 0, 0, 0);
+    returnQuartelRoomL->collisionEnabled = 1;
+    returnQuartelRoomL->collisionType = COLLISION_RECT;
+}
